DynamoDB Item Validation with AST and Visitor Pattern
This works great in Python 3, but I need to make major
changes to get it to work on Python 2/3.

Still need to clean up some bits, especially those that
are obviously only relevant to the original GraphQL source,
e.g. all the Graphql query nodes in the QUERY_DOCUMENT_KEYS

Had to make minor modifications to a couple a tests that were
incorrectly mixing numbers and strings.