Adding a little black magic to the serializer
For some attributes, they have a prefix that's a bit redundant, like ClusterCreateTime

So now the serializer will look for a variety of possible attributes that match that:

ClusterCreateTime
cluster_create_time
class Cluster().CreateTime
class Cluster().create_time

This is also helpful at the moment because there is a lot of inconsistencies among the
moto models.  Object keys are useful snake case, but a lot of dicts have the Pascal casing.