Fix Input Parsing
A recent version of botocore added a new parameter, `Engine`, to the ModifyDBInstance
input.  We were already passing in `EngineVersion`, so the old `_has_member` method
was picking up that `Engine` was in the input (because `i.startswith("Engine")` was
True.

I think this code originally was meant to help deal with input that had multiple
values, e.g. VPCSecurityId.1, etc. but I'm no longer sure that it was necessary
as originally written.  It was definitely incorrect though, but only got exposed
today when the data models changed in a version specific way.

In any case, I've made the change to do a more direct comparison, as well as
a comparison for any `.` values that *may* exist.

All RDS tests (for our moto fork) were passing both before and after this change,
but this change should resolve the current issue with the RDS Broker, where two
tests are failing (that call ModifyDBInstance).

Ref: https://github.com/getmoto/moto/security/code-scanning/76