Migrate SetRulePrioritiesTemplate and Create/Describe/ModifyRule(s)
This one required bringing in the QueryParser from RDS (and fixing a bug
there and in the get_value method).
The basic issue that had to be solved is that the rules rely on some "map" objects
and the current moto parameter parser did not handle them properly, which propagated
them to the backend and caused issue with serialization.