Default Value Problem
This needs some thinking.  Where do we handle default values?  In responses or in the model?  Futhermore,
the parser doesn't return values that are None.  Like, it is important to distinguish between a value not
being provided and a null value specified.  I feel like this is something that Pydantic and Marshmallow
and others deal with, but I need to look into it more.  (Also check Django)